+++ explabox/digestibles.py
@@ -1,6 +1,7 @@
 """Ingestibles are turned into digestibles, containing information to explore/examine/explain/expose your model."""
 
 from genbase import MetaInfo
+from genbase.utils import extract_metrics, recursive_to_dict
 
 from .ui.notebook import Render
 
@@ -10,15 +11,16 @@ class Performance(MetaInfo):
         """..."""
         super().__init__(type=type, subtype=subtype, renderer=Render, **kwargs)
         self.labels = labels
-        self._metrics = metrics
+        self._metrics, self._properties = extract_metrics(metrics)
 
     @property
     def metrics(self):
-        return self._metrics.__dict__
+        return self._metrics
 
     @property
     def content(self):
-        return {"labels": self.labels, "metrics": self.metrics}
+        label_metrics = [{"label": label, "metrics": self.metrics[label]} for label in self.labels]
+        return {"labels": self.labels, "label_metrics": label_metrics, "metrics": self._properties}
 
 
 class Descriptives(MetaInfo):