+++ explabox/expose/text/__init__.py
@@ -1,11 +1,17 @@
 """Functions/classes for sensitivity testing (fairness and robustness) for text data."""
 
+from typing import List, Optional, Union
+
 from genbase import Readable
 from text_sensitivity import (OneToManyPerturbation, OneToOnePerturbation,
-                              compare_accuracy, compare_metric,
-                              compare_precision, compare_recall,
-                              input_space_robustness, invariance, mean_score,
-                              perturbation)
+                              RandomAscii, RandomCyrillic, RandomDigits,
+                              RandomEmojis, RandomLower, RandomPunctuation,
+                              RandomSpaces, RandomString, RandomUpper,
+                              RandomWhitespace, compare_accuracy,
+                              compare_metric, compare_precision,
+                              compare_recall, input_space_robustness,
+                              invariance, mean_score, perturbation)
+from text_sensitivity.return_types import SuccessTest
 
 from ...ingestibles import Ingestible
 from ...mixins import IngestiblesMixin
@@ -22,4 +28,35 @@ class Exposer(Readable, IngestiblesMixin):
         if ingestibles is None:
             ingestibles = Ingestible(data=data, model=model)
         self.ingestibles = ingestibles
-        self.check_requirements(['data', 'model'])
+        self.check_requirements(["data", "model"])
+
+    def input_space(
+        self,
+        generators: List[Union[RandomString, str]],
+        n_samples: int = 100,
+        min_length: int = 0,
+        max_length: int = 100,
+        seed: Optional[int] = 0,
+        **kwargs
+    ) -> SuccessTest:
+        GENERATORS = {
+            "ascii": RandomAscii(),
+            "emojis": RandomEmojis(),
+            "whitespaces": RandomWhitespace(),
+            "spaces": RandomSpaces(),
+            "ascii_upper": RandomUpper(),
+            "acii_lower": RandomLower(),
+            "digits": RandomDigits(),
+            "punctuation": RandomPunctuation(),
+            "cyrillic": RandomCyrillic(),
+        }
+        generator = [GENERATORS[generator] if isinstance(generator, str) and generator in GENERATORS else generator]
+        return input_space_robustness(
+            model=self.model,
+            generators=generators,
+            n_samples=n_samples,
+            min_length=min_length,
+            max_length=max_length,
+            seed=seed,
+            **kwargs
+        )