+++ explabox/ingestibles/__init__.py
@@ -7,28 +7,30 @@ from instancelib import InstanceProvider, MemoryEnvironment
 from instancelib.typehints import KT
 from sklearn.exceptions import NotFittedError
 
-from .data import import_data, train_test_split
+from .data import import_data, rename_labels, train_test_split
 from .model import import_model
 
 
 class Ingestible(dict):
-    def __init__(self,
-                 data=None,
-                 model=None,
-                 splits: Dict[KT, KT] = {'train': 'train', 'test': 'test', 'validation': 'validation'}):
-        self['data'] = data
-        self['model'] = model
+    def __init__(
+        self,
+        data=None,
+        model=None,
+        splits: Dict[KT, KT] = {"train": "train", "test": "test", "validation": "validation"},
+    ):
+        self["data"] = data
+        self["model"] = model
         self.__splits = splits
 
     @property
     def environment(self):
-        if hasattr(self, '_environment'):
+        if hasattr(self, "_environment"):
             return self._environment
         return MemoryEnvironment(self.data, self.labelprovider)
 
     @property
     def data(self):
-        return self['data']
+        return self["data"]
 
     def get_named_split(self, name: KT) -> Optional[InstanceProvider]:
         if name in self.__splits.keys() and self.__splits[name] in self.data.keys():
@@ -39,15 +41,15 @@ class Ingestible(dict):
 
     @property
     def train(self):
-        return self.get_named_split('train')
+        return self.get_named_split("train")
 
     @property
     def test(self):
-        return self.get_named_split('test')
+        return self.get_named_split("test")
 
     @property
     def validation(self):
-        return self.get_named_split('validation')
+        return self.get_named_split("validation")
 
     @property
     def splits(self):
@@ -67,13 +69,13 @@ class Ingestible(dict):
 
     @property
     def model(self):
-        return self['model']
+        return self["model"]
 
     @model.setter
     def model(self, model):
-        self['model'] = model
+        self["model"] = model
 
-    def check_requirements(self, elements: List[str] = ['data', 'model']):
+    def check_requirements(self, elements: List[str] = ["data", "model"]):
         """Check if the required elements are in the ingestibles.
 
         Args: