+++ explabox/ingestibles/data/__init__.py
@@ -1,3 +1,3 @@
 """Handling of data."""
 
-from genbase.data import import_data, train_test_split
+from genbase.data import import_data, rename_labels, train_test_split