+++ docs/source/changelog.rst
@@ -22,8 +22,8 @@ Added
 * Dataset descriptives
 * Text sensitivity tests (\ ``text_sensitivity``\ )
 * Text explainability (\ ``text_explainability``\ )
-* Model importing (\ ``instancelib``\ )
-* Dataset handling (\ ``instancelib``\ )
+* Model importing (\ ``genbase``\ )
+* Dataset handling (\ ``genbase``\ )
 * Basic UI (\ ``genbase``\ )
 * ``git`` setup